Remove protobuf version restriction in conda recipe (#10159)
## Describe your changes

We restricted protobuf due to issues in conda, but we have since
upgraded protobuf. We should have it mimic reality. For this, we should
try to not get involved and let the setup.py dictate. Assuming the Conda
build succeeds, then new versions should be better and we can undo the
check.

## Testing Plan

- Conda Build should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.