Consolidate border radii value and remove redundant css attr (#10184)
## Describe your changes

1. Move border radii (xxxl) to const so we have 1 source of truth for
its value throughout this component. In the future when we want to
change it to consume advanced theming's config we only need to modify 1
const.
2. There are quite a few places where we set `backgroundColor:
theme.colors.transparent` but they did not change the rendered outcome,
or could be handled by setting `overflow: "hidden"` in their parent. So
I re-organized the css attributes in these parent-children element and
a) move the `focus` attr to `StyledChatInputContainer` and b) remove
redundant attr after adding `overflow: hidden`


## GitHub Issue Link (if applicable)

## Testing Plan

Manual tested that chat input w or w/o file upload behaves the same

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.