Change dataframe column format from UI (#10420)
## Describe your changes

Adds a new option to the dataframe column menu that allows users to
change the value formatting of the column to one of the common
pre-defined formats. This is available for all columns that support the
`format` paramter (number, progress, datetime, date, & time).


https://github.com/user-attachments/assets/ac36475f-6879-4355-ac14-d992704ab807

## Testing Plan

- Added unit tests.
- Added e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.