Activate Pandas linting rules (PD) (#10571)
## Describe your changes

Activates the pandas linting category in ruff (PD). This provides a
couple of opinionated linting rules for usage of the Pandas API. We
ignore this rule in our tests since our tests do not need the same
scrutiny.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.