Add test execution time to e2e performance tracking (#10635)
## Describe your changes

Adds a custom metric to our performance tracking that captures the total
execution time of the test logic. While it doesn't perfectly fit in with
the existing browser based metrics, it seems like the easiest way to
track this execution number for some of our tests.

This is mainly relevant for some follow-up performance tracking related
to the forward message cache:
https://github.com/streamlit/streamlit/pull/10638

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.