Make test get_min_versions script works with nightly builds (#10646)
## Describe your changes

This script only looks for the `streamlit` package, but now that we are
using it in the nightly build pipeline we should also look for the
`streamlit-nightly` package if we can't find the `streamlit` pacakge.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.